requirements.txt
ai21==4.0.3
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
distro==1.9.0
Django==5.2.4
docutils==0.21.2
exceptiongroup==1.3.0
filetype==1.2.0
Flask==2.3.3
flask-cors==6.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
openai==1.91.0
packaging==25.0
pillow==11.2.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-decouple==3.8
python-dotenv==1.1.1
pyttsx3==2.98
pytweening==1.2.0
pywhatkit==5.4
pywin32==310
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
sqlparse==0.5.3
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.9.0
wikipedia==1.4.0
